Tests: more timeouts
